automated_defect_detection/image_processor.py
import cv2
import numpy as np
from models import Defect

class ObjectDetectionModel:
    def __init__(self, model_path):
        self.modelPath = model_path
        self.model = None

    def load(self):
        try:
            # Load your YOLO model here
            # self.model = cv2.dnn.readNet(self.modelPath)
            print("YOLO model loaded successfully.")
            return True
        except Exception as e:
            print(f"Error loading model: {e}")
            return False

    def predict(self, image_file):
        # Placeholder for object detection logic
        # Returns a list of dummy defects
        return [
            {"type": "crack", "boundingBox": [100, 50, 250, 180], "confidence": 0.92},
            {"type": "dent", "boundingBox": [300, 200, 400, 300], "confidence": 0.85}
        ]

class ImageProcessor:
    def __init__(self):
        self.opencvLib = cv2
        self.numpyLib = np
        self.yoloModel = ObjectDetectionModel("path/to/yolov4-tiny.weights")
        self.yoloModel.load()

    def preprocess(self, image_path):
        # Image preprocessing logic
        image = self.opencvLib.imread(image_path)
        gray_image = self.opencvLib.cvtColor(image, self.opencvLib.COLOR_BGR2GRAY)
        # Add more preprocessing steps like resizing, noise removal etc.
        return gray_image

    def detect(self, image_file):
        # Use the YOLO model to detect defects
        detected_defects_data = self.yoloModel.predict(image_file)
        defects = [Defect(image_file.imageID, d['type'], d['boundingBox'], d['confidence']) for d in detected_defects_data]
        return defects

    def highlight(self, image_file, defects):
        # Draw bounding boxes and labels on the image
        image = self.opencvLib.imread(image_file.originalPath)
        for defect in defects:
            x1, y1, x2, y2 = defect.boundingBox
            self.opencvLib.rectangle(image, (x1, y1), (x2, y2), (0, 255, 0), 2)
            self.opencvLib.putText(image, defect.type, (x1, y1 - 10), self.opencvLib.FONT_HERSHEY_SIMPLEX, 0.9, (0, 255, 0), 2)
        
        # Save the processed image to a new path
        processed_path = f"processed_images/{image_file.imageID}_processed.jpg"
        self.opencvLib.imwrite(processed_path, image)
        return processed_path