requirements.txt
customtkinter
Pillow
mysql-connector-python
numpy
opencv-python
